requirements.txt
streamlit
scikit-learn
opencv-python-headless
pillow
numpy